ver 1
ZIG Telematics offers an advanced solution that combines hardware and software for vehicle tracking. Our state-of-the-art Web and Mobile app-based tracking system provides real-time monitoring of your vehicles and sends on-demand alerts and notifications.

Our tracking system is designed to meet all your tracking needs and is user-friendly with an intuitive design. Additionally, we offer unique features that are not available in other tracking systems. You can get all these benefits at an unmatched price in the market.

https://play.google.com/store/apps/details?id=com.zig.gps&hl=en_IN&gl=US